관리자 api 완성

+++ backend/cinemaseoul/src/main/java/com/uos/cinemaseoul/controller/user/AdminController.java
@@ -3,6 +3,7 @@ package com.uos.cinemaseoul.controller.user;
 import com.uos.cinemaseoul.common.auth.AuthUser;
 import com.uos.cinemaseoul.common.auth.JwtTokenProvider;
 import com.uos.cinemaseoul.common.auth.UserType;
+import com.uos.cinemaseoul.common.paging.Criteria;
 import com.uos.cinemaseoul.dto.user.*;
 import com.uos.cinemaseoul.exception.BlackListException;
 import com.uos.cinemaseoul.exception.DuplicateException;
@@ -14,6 +15,7 @@ import com.uos.cinemaseoul.vo.user.BlackListVo;
 import com.uos.cinemaseoul.vo.user.UsersVo;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.Authentication;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.web.bind.annotation.*;
 
@@ -35,7 +37,7 @@ public class AdminController {
         }else if(!passwordEncoder.matches(logindto.getPassword(), authUser.getPassword())){
             throw new WrongArgException("Wrong Password");
         }
-        return jwtTokenProvider.createToken(UserType.USERS.toString(), authUser.getUsername(), authUser.getAuth());
+        return jwtTokenProvider.createToken(UserType.ADMIN.toString(), authUser.getUsername(), authUser.getAuth());
     }
 
     @PostMapping("/signup")
@@ -59,9 +61,14 @@ public class AdminController {
 
     //관리자 수정
     @PutMapping("")
-    public void updateAdmi(@RequestBody final AdminVo adv)throws Exception{
+    public void updateAdmi(Authentication authentication, @RequestBody final AdminVo adv)throws Exception{
+
+        //매니저도 아니고, 자기 id도 아니면
+        if(!authentication.getAuthorities().contains("ROLE_3")
+                && Integer.parseInt(authentication.getName()) != adv.getAdmi_id()) throw new DuplicateException();
+
         try{
-            int result = adminService.updateAdmi(adv);
+            int result = adminService.updateAdmin(adv);
             if(result == 0){
                 throw new WrongArgException("Wrong Update Form");
             }
@@ -71,9 +78,22 @@ public class AdminController {
         }
     }
 
-    //관리자 조회
+    //리스트 조회
+    @PostMapping("/list")
+    public ResponseEntity<AdminListDto> selectList(@RequestBody Criteria criteria){
+        return ResponseEntity.ok(adminService.selectList(criteria));
+
+    }
+
+    //관리자 세부 조회
     @GetMapping("/{admi_id}")
-    public ResponseEntity<AdminInfoDto> selectById(@PathVariable int admi_id){
+    public ResponseEntity<AdminInfoDto> selectById(Authentication authentication, @PathVariable int admi_id){
+
+        //매니저도 아니고, 자기 id도 아니면
+        if(!authentication.getAuthorities().contains("ROLE_3")
+                && Integer.parseInt(authentication.getName()) != admi_id ) throw new DuplicateException();
+
+
         AdminInfoDto adminInfoDto;
         try{
             adminInfoDto = adminService.selectById(admi_id);
@@ -89,7 +109,11 @@ public class AdminController {
 
     //관리자탈퇴
     @DeleteMapping("/{admi_id}")
-    public void deleteUser(@PathVariable int admi_id)throws Exception{
+    public void deleteUser(Authentication authentication, @PathVariable int admi_id)throws Exception{
+        //매니저도 아니고, 자기 id도 아니면
+        if(!authentication.getAuthorities().contains("ROLE_3")
+                && Integer.parseInt(authentication.getName()) != admi_id ) throw new DuplicateException();
+
         try {
             int result = adminService.deleteAdmin(admi_id);
             if(result == 0){
@@ -103,7 +127,7 @@ public class AdminController {
 
     //번호검사
     @PostMapping("/phonecheck")
-    public void phoneCheck(@RequestParam(name = "phone_num") String phone_num){
+    public void phoneCheck(@ModelAttribute(name = "phone_num") String phone_num){
         if(!adminService.phoneCheck(phone_num)){
             throw new DuplicateException();
         }
@@ -111,7 +135,7 @@ public class AdminController {
 
     //이메일검사
     @PostMapping("/emailcheck")
-    public void emailCheck(@RequestParam(name = "email") String email){
+    public void emailCheck(@ModelAttribute(name = "email") String email){
         if(!adminService.emailCheck(email)){
             throw new DuplicateException();
         }
